Enhance CMake configuration and add GPU kernel benchmarking tool
- Disable automatic dependency deployment in vcpkg.
- Improve FFTW3 detection messages for better clarity.
- Update various functions to suppress unused parameter warnings.
- Implement GPU kernel performance benchmarking tool for envelope detection and phase unwrapping.
- Modify viewer build commands for better compatibility on Windows.
- Adjust texture handling in VideoWidget for correct image orientation.

+++ cpp-prototype/src/demod/fm_demodulator.cpp
@@ -145,7 +145,7 @@ FMDemodulator::Result FMDemodulator::demodulate(const ComplexArray& analyticSign
     Result result;
     
     // FM demodulation using Python-compatible algorithm
-    float sampleRate = config_.inputFreqMHz * 1e6f;  // Convert MHz to Hz
+    float sampleRate = static_cast<float>(config_.inputFreqMHz * 1e6);  // Convert MHz to Hz
     result.video = impl_->fmDemodulate(analyticSignal, sampleRate);
     
     // Detect envelope for dropout detection
